refactor!: Applying changes from standalone branch
Changes:
- Refactored packages: Coding & Unitizing studies are now in their own
packages, as are serialization classes, TTLab special classes and the
new relational agreement classes.
- Added classes for relational agreement calculation.
- Updated dependencies.

All tests pass.

BREAKING CHANGE: *All* classes are now in different packages.
