+++ pom.xml
@@ -91,15 +91,15 @@
         </dependency>
         <dependency>
             <groupId>org.dkpro.core</groupId>
+            <artifactId>dkpro-core-api-segmentation-asl</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.dkpro.core</groupId>
             <artifactId>dkpro-core-api-metadata-asl</artifactId>
         </dependency>
         <dependency>
             <groupId>org.dkpro.core</groupId>
-            <artifactId>dkpro-core-api-parameter-asl</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.dkpro.core</groupId>
-            <artifactId>dkpro-core-io-xmi-asl</artifactId>
+            <artifactId>dkpro-core-api-io-asl</artifactId>
         </dependency>
         <dependency>
             <groupId>org.json</groupId>
