+++ src/main/java/org/hucompute/textimager/uima/agreement/engine/coding/SetSelectionStrategy.java
@@ -1,0 +1,5 @@
+package org.hucompute.textimager.uima.agreement.engine.coding;
+
+public enum SetSelectionStrategy {
+	MAX, ALL, MATCH
+}
