+++ src/main/java/org/hucompute/textimager/uima/agreement/package-info.java
@@ -1,0 +1,14 @@
+/**
+ * biofid.agreement is used to calculate the inter-annotator agreement for annotations in an UIMA JCas and is
+ * part of the TextAnnotator & TextImager pipelines.<br>
+ * It was created by the Text Technology Lab, Goethe University Frankfurt for BIOfid project.<br>
+ * <p/>
+ * Author: Manuel Stoeckel<br>
+ * Contact: manuel.stoeckel (at) stud.uni-frankfurt.de<br>
+ * Licence: AGPL-3.0<br>
+ * @see <a href="https://www.biofid.de/">BIOfid</a>
+ * @see <a href="https://www.texttechnologylab.org/">Text Technology Lab</a>
+ * @see <a href="https://textimager.hucompute.org/">TextImager</a>
+ * @see <a href="http://www.textannotator.texttechnologylab.org/">TextAnnotator</a>
+ */
+package org.hucompute.textimager.uima.agreement;
