--- src/main/java/org/texttechnologylab/agreement/engine/SetSelectionStrategy.java
@@ -1,5 +1,0 @@
-package org.texttechnologylab.agreement.engine;
-
-public enum SetSelectionStrategy {
-    MAX, ALL, MATCH
-}
