+++ src/test/resources/META-INF/org.apache.uima.fit/types.txt
