+++ src/test/resources/test-Anne_Frank.xmi
